backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.6
websockets==12.0
openai==1.3.7
pillow==10.1.0
numpy==1.26.2
# Memory and Task Services
croniter==2.0.1
# Discord Bot
discord.py==2.3.2
# Telegram Bot
python-telegram-bot==20.7
# WhatsApp Bot (optional - requires Twilio)
twilio==8.10.0
